requirements.txt
Levenshtein==0.21.0
mysql-connector-python==8.0.33
protobuf==3.20.3
rapidfuzz==3.0.0
xmltodict==0.13.0
